dom/base/test/test_bug614058.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=614058
-->
<head>
  <title>Test for Bug 614058</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=614058">Mozilla Bug 614058</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 614058 **/
var f1 = document.createDocumentFragment(); 
f2 = f1.cloneNode(true); 
f1.appendChild(document.createElement("foo"));
is(f1.isEqualNode(f2), false, "Fragments have different kids!");

</script>
</pre>
</body>
</html>
